Issue #44 iterate kokkos grids

+++ src/datastream.hpp
@@ -38,8 +38,8 @@
  */
 template<
   class T, 
-  int length, 
-  template<int> class CheckingPolicy
+  size_t length, 
+  template<size_t> class CheckingPolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Array<T,length,CheckingPolicy>&);
 
@@ -49,7 +49,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Array<T,length,CheckingPol
  */
 template<
   typename T,
-  template<int> class CheckingPolicy,
+  template<size_t> class CheckingPolicy,
   template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 1, CheckingPolicy, StoragePolicy>&);
@@ -61,7 +61,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 1, CheckingPolicy,
  */
 template<
   typename T,
-  template<int> class CheckingPolicy,
+  template<size_t> class CheckingPolicy,
   template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 2, CheckingPolicy, StoragePolicy>&);
@@ -74,7 +74,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 2, CheckingPolicy,
  */
 template<
   typename T,
-  template<int> class CheckingPolicy,
+  template<size_t> class CheckingPolicy,
   template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 3, CheckingPolicy, StoragePolicy>&);
@@ -87,7 +87,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 3, CheckingPolicy,
  */
 template<
   typename T,
-  template<int> class CheckingPolicy,
+  template<size_t> class CheckingPolicy,
   template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 4, CheckingPolicy, StoragePolicy>&);
@@ -101,7 +101,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 4, CheckingPolicy,
  */
 template<
   typename T,
-  template<int> class CheckingPolicy,
+  template<size_t> class CheckingPolicy,
   template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 5, CheckingPolicy, StoragePolicy>&);
